refactor(style): use tailwind base and scope Preflight
Before, `@tailwind base` was not applied by default because article pages don't use Preflight, but it was causing lots of issues:

> In Tailwind CSS v3.0, the `@tailwind base` directive must be present for utilities like transforms, filters, and shadows to work as expected. https://tailwindcss.com/docs/upgrade-guide#base-layer-must-be-present

This commit applies `@tailwind base` everywhere and simplifies the styling of the navbar, transforms, etc. accordingly. It also adds a plugin to scope Preflight, or in my case, exclude it from certain divs.